requirements.txt
discord.py==2.1.0
mysql-connector-python==8.0.32
colorama==0.4.6
emoji==2.2.0
python-a2s==1.3.0
pynacl==1.5.0
steam==1.4.4
gevent==22.10.2
gevent-eventemitter==2.1